+++ pkg/deployment/resources/pod_creator.go
@@ -17,8 +17,6 @@
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
-// Author Ewout Prangsma
-// Author Tomasz Mielech
 //
 
 package resources
@@ -35,6 +33,12 @@ import (
 	"sync"
 	"time"
 
+	podMod "github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/pod"
+	core "k8s.io/api/core/v1"
+	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/tls"
+
 	"github.com/arangodb/kube-arangodb/pkg/util"
 
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
@@ -44,18 +48,13 @@ import (
 	inspectorInterface "github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/interfaces"
 
+	"github.com/arangodb/go-driver"
 	"k8s.io/apimachinery/pkg/types"
 
-	"github.com/arangodb/kube-arangodb/pkg/deployment/pod"
-
-	driver "github.com/arangodb/go-driver"
 	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
+	"github.com/arangodb/kube-arangodb/pkg/deployment/pod"
 	"github.com/arangodb/kube-arangodb/pkg/util/constants"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
-
-	core "k8s.io/api/core/v1"
-	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
 )
 
 func versionHasAdvertisedEndpoint(v driver.Version) bool {
@@ -350,10 +349,6 @@ func (r *Resources) RenderPodForMember(ctx context.Context, cachedStatus inspect
 	}
 	groupSpec := spec.GetServerGroupSpec(group)
 
-	kubecli := r.context.GetKubeCli()
-	ns := r.context.GetNamespace()
-	secrets := kubecli.CoreV1().Secrets(ns)
-
 	memberName := m.ArangoMemberName(r.context.GetAPIObject().GetName(), group)
 
 	member, ok := cachedStatus.ArangoMember(memberName)
@@ -369,7 +364,8 @@ func (r *Resources) RenderPodForMember(ctx context.Context, cachedStatus inspect
 
 	newMember.PodName = k8sutil.CreatePodName(apiObject.GetName(), roleAbbr, newMember.ID, CreatePodSuffix(spec))
 
-	// Render pod
+	var podCreator interfaces.PodCreator
+	var args []string
 	if group.IsArangod() {
 		// Prepare arguments
 		autoUpgrade := newMember.Conditions.IsTrue(api.ConditionTypeAutoUpgrade) || spec.Upgrade.Get().AutoUpgrade
@@ -390,7 +386,8 @@ func (r *Resources) RenderPodForMember(ctx context.Context, cachedStatus inspect
 
 		input := memberPod.AsInput()
 
-		args, err := createArangodArgs(cachedStatus, input)
+		var err error
+		args, err = createArangodArgs(cachedStatus, input)
 		if err != nil {
 			return nil, errors.WithStack(err)
 		}
@@ -398,8 +395,7 @@ func (r *Resources) RenderPodForMember(ctx context.Context, cachedStatus inspect
 		if err := memberPod.Validate(cachedStatus); err != nil {
 			return nil, errors.WithStack(errors.Wrapf(err, "Validation of pods resources failed"))
 		}
-
-		return RenderArangoPod(cachedStatus, apiObject, role, newMember.ID, newMember.PodName, args, &memberPod)
+		podCreator = &memberPod
 	} else if group.IsArangosync() {
 		// Check image
 		if !imageInfo.Enterprise {
@@ -417,7 +413,7 @@ func (r *Resources) RenderPodForMember(ctx context.Context, cachedStatus inspect
 
 		masterJWTSecretName = spec.Sync.Authentication.GetJWTSecretName()
 		err := k8sutil.RunWithTimeout(ctx, func(ctxChild context.Context) error {
-			return k8sutil.ValidateTokenSecret(ctxChild, secrets, masterJWTSecretName)
+			return k8sutil.ValidateTokenSecret(ctxChild, cachedStatus.SecretReadInterface(), masterJWTSecretName)
 		})
 		if err != nil {
 			return nil, errors.WithStack(errors.Wrapf(err, "Master JWT secret validation failed"))
@@ -425,7 +421,7 @@ func (r *Resources) RenderPodForMember(ctx context.Context, cachedStatus inspect
 
 		monitoringTokenSecretName := spec.Sync.Monitoring.GetTokenSecretName()
 		err = k8sutil.RunWithTimeout(ctx, func(ctxChild context.Context) error {
-			return k8sutil.ValidateTokenSecret(ctxChild, secrets, monitoringTokenSecretName)
+			return k8sutil.ValidateTokenSecret(ctxChild, cachedStatus.SecretReadInterface(), monitoringTokenSecretName)
 		})
 		if err != nil {
 			return nil, errors.WithStack(errors.Wrapf(err, "Monitoring token secret validation failed"))
@@ -438,7 +434,7 @@ func (r *Resources) RenderPodForMember(ctx context.Context, cachedStatus inspect
 			if spec.IsAuthenticated() {
 				clusterJWTSecretName = spec.Authentication.GetJWTSecretName()
 				err = k8sutil.RunWithTimeout(ctx, func(ctxChild context.Context) error {
-					return k8sutil.ValidateTokenSecret(ctxChild, secrets, clusterJWTSecretName)
+					return k8sutil.ValidateTokenSecret(ctxChild, cachedStatus.SecretReadInterface(), clusterJWTSecretName)
 				})
 				if err != nil {
 					return nil, errors.WithStack(errors.Wrapf(err, "Cluster JWT secret validation failed"))
@@ -447,7 +443,7 @@ func (r *Resources) RenderPodForMember(ctx context.Context, cachedStatus inspect
 			// Check client-auth CA certificate secret
 			clientAuthCASecretName = spec.Sync.Authentication.GetClientCASecretName()
 			err = k8sutil.RunWithTimeout(ctx, func(ctxChild context.Context) error {
-				return k8sutil.ValidateCACertificateSecret(ctxChild, secrets, clientAuthCASecretName)
+				return k8sutil.ValidateCACertificateSecret(ctxChild, cachedStatus.SecretReadInterface(), clientAuthCASecretName)
 			})
 			if err != nil {
 				return nil, errors.WithStack(errors.Wrapf(err, "Client authentication CA certificate secret validation failed"))
@@ -455,7 +451,7 @@ func (r *Resources) RenderPodForMember(ctx context.Context, cachedStatus inspect
 		}
 
 		// Prepare arguments
-		args := createArangoSyncArgs(apiObject, spec, group, groupSpec, *newMember)
+		args = createArangoSyncArgs(apiObject, spec, group, groupSpec, *newMember)
 
 		memberSyncPod := MemberSyncPod{
 			tlsKeyfileSecretName:   tlsKeyfileSecretName,
@@ -470,10 +466,23 @@ func (r *Resources) RenderPodForMember(ctx context.Context, cachedStatus inspect
 			arangoMember:           *member,
 		}
 
-		return RenderArangoPod(cachedStatus, apiObject, role, newMember.ID, newMember.PodName, args, &memberSyncPod)
+		podCreator = &memberSyncPod
 	} else {
 		return nil, errors.Newf("unable to render Pod")
 	}
+
+	pod, err := RenderArangoPod(cachedStatus, apiObject, role, newMember.ID, newMember.PodName, args, podCreator)
+	if err != nil {
+		return nil, err
+	}
+
+	if features.RandomPodNames().Enabled() {
+		// The server will generate the name with some additional suffix after `-`.
+		pod.GenerateName = pod.Name + "-"
+		pod.Name = ""
+	}
+
+	return pod, nil
 }
 
 func (r *Resources) SelectImage(spec api.DeploymentSpec, status api.DeploymentStatus) (api.ImageInfo, bool) {
@@ -503,7 +512,7 @@ func (r *Resources) SelectImageForMember(spec api.DeploymentSpec, status api.Dep
 }
 
 // createPodForMember creates all Pods listed in member status
-func (r *Resources) createPodForMember(ctx context.Context, spec api.DeploymentSpec, member *api.ArangoMember, memberID string, imageNotFoundOnce *sync.Once) error {
+func (r *Resources) createPodForMember(ctx context.Context, cachedStatus inspectorInterface.Inspector, spec api.DeploymentSpec, member *api.ArangoMember, memberID string, imageNotFoundOnce *sync.Once) error {
 	log := r.log
 	status, lastVersion := r.context.GetStatus()
 
@@ -538,11 +547,8 @@ func (r *Resources) createPodForMember(ctx context.Context, spec api.DeploymentS
 
 	imageInfo = *m.Image
 
-	kubecli := r.context.GetKubeCli()
 	apiObject := r.context.GetAPIObject()
 
-	ns := r.context.GetNamespace()
-	secrets := kubecli.CoreV1().Secrets(ns)
 	if !found {
 		return errors.WithStack(errors.Newf("Member '%s' not found", memberID))
 	}
@@ -563,11 +569,12 @@ func (r *Resources) createPodForMember(ctx context.Context, spec api.DeploymentS
 
 		ctxChild, cancel := context.WithTimeout(ctx, k8sutil.GetRequestTimeout())
 		defer cancel()
-		uid, err := CreateArangoPod(ctxChild, kubecli, apiObject, spec, group, CreatePodFromTemplate(template.PodSpec))
+		podName, uid, err := CreateArangoPod(ctxChild, r.context.PodsModInterface(), apiObject, spec, group, CreatePodFromTemplate(template.PodSpec))
 		if err != nil {
 			return errors.WithStack(err)
 		}
 
+		m.PodName = podName
 		m.PodUID = uid
 		m.PodSpecVersion = template.PodSpecChecksum
 		m.ArangoVersion = m.Image.ArangoDBVersion
@@ -590,19 +597,25 @@ func (r *Resources) createPodForMember(ctx context.Context, spec api.DeploymentS
 		if group == api.ServerGroupSyncMasters {
 			// Create TLS secret
 			tlsKeyfileSecretName := k8sutil.CreateTLSKeyfileSecretName(apiObject.GetName(), role, m.ID)
-			serverNames := []string{
+
+			names, err := tls.GetAltNames(spec.Sync.TLS)
+			if err != nil {
+				return errors.WithStack(errors.Wrapf(err, "Failed to render alt names"))
+			}
+
+			names.AltNames = append(names.AltNames,
 				k8sutil.CreateSyncMasterClientServiceName(apiObject.GetName()),
 				k8sutil.CreateSyncMasterClientServiceDNSNameWithDomain(apiObject, spec.ClusterDomain),
 				k8sutil.CreatePodDNSNameWithDomain(apiObject, spec.ClusterDomain, role, m.ID),
-			}
+			)
 			masterEndpoint := spec.Sync.ExternalAccess.ResolveMasterEndpoint(k8sutil.CreateSyncMasterClientServiceDNSNameWithDomain(apiObject, spec.ClusterDomain), k8sutil.ArangoSyncMasterPort)
 			for _, ep := range masterEndpoint {
 				if u, err := url.Parse(ep); err == nil {
-					serverNames = append(serverNames, u.Hostname())
+					names.AltNames = append(names.AltNames, u.Hostname())
 				}
 			}
 			owner := apiObject.AsOwner()
-			_, err := createTLSServerCertificate(ctx, log, secrets, serverNames, spec.Sync.TLS, tlsKeyfileSecretName, &owner)
+			_, err = createTLSServerCertificate(ctx, log, cachedStatus, r.context.SecretsModInterface(), names, spec.Sync.TLS, tlsKeyfileSecretName, &owner)
 			if err != nil && !k8sutil.IsAlreadyExists(err) {
 				return errors.WithStack(errors.Wrapf(err, "Failed to create TLS keyfile secret"))
 			}
@@ -610,12 +623,13 @@ func (r *Resources) createPodForMember(ctx context.Context, spec api.DeploymentS
 
 		ctxChild, cancel := context.WithTimeout(ctx, k8sutil.GetRequestTimeout())
 		defer cancel()
-		uid, err := CreateArangoPod(ctxChild, kubecli, apiObject, spec, group, CreatePodFromTemplate(template.PodSpec))
+		podName, uid, err := CreateArangoPod(ctxChild, r.context.PodsModInterface(), apiObject, spec, group, CreatePodFromTemplate(template.PodSpec))
 		if err != nil {
 			return errors.WithStack(err)
 		}
 		log.Debug().Str("pod-name", m.PodName).Msg("Created pod")
 
+		m.PodName = podName
 		m.PodUID = uid
 		m.PodSpecVersion = template.PodSpecChecksum
 	}
@@ -704,13 +718,14 @@ func RenderArangoPod(cachedStatus inspectorInterface.Inspector, deployment k8sut
 // CreateArangoPod creates a new Pod with container provided by parameter 'containerCreator'
 // If the pod already exists, nil is returned.
 // If another error occurs, that error is returned.
-func CreateArangoPod(ctx context.Context, kubecli kubernetes.Interface, deployment k8sutil.APIObject, deploymentSpec api.DeploymentSpec, group api.ServerGroup, pod *core.Pod) (types.UID, error) {
-	uid, err := k8sutil.CreatePod(ctx, kubecli, pod, deployment.GetNamespace(), deployment.AsOwner())
+func CreateArangoPod(ctx context.Context, c podMod.ModInterface, deployment k8sutil.APIObject,
+	deploymentSpec api.DeploymentSpec, group api.ServerGroup, pod *core.Pod) (string, types.UID, error) {
+	podName, uid, err := k8sutil.CreatePod(ctx, c, pod, deployment.GetNamespace(), deployment.AsOwner())
 	if err != nil {
-		return "", errors.WithStack(err)
+		return "", "", errors.WithStack(err)
 	}
 
-	return uid, nil
+	return podName, uid, nil
 }
 
 func CreatePodFromTemplate(p *core.PodTemplateSpec) *core.Pod {
@@ -764,7 +779,7 @@ func (r *Resources) EnsurePods(ctx context.Context, cachedStatus inspectorInterf
 			r.log.Warn().Msgf("Ensuring pod")
 
 			spec := r.context.GetSpec()
-			if err := r.createPodForMember(ctx, spec, member, m.ID, imageNotFoundOnce); err != nil {
+			if err := r.createPodForMember(ctx, cachedStatus, spec, member, m.ID, imageNotFoundOnce); err != nil {
 				r.log.Warn().Err(err).Msgf("Ensuring pod failed")
 				return errors.WithStack(err)
 			}
@@ -777,8 +792,11 @@ func (r *Resources) EnsurePods(ctx context.Context, cachedStatus inspectorInterf
 	return nil
 }
 
+// CreatePodSuffix creates additional string to glue it to the POD name.
+// The suffix is calculated according to the given spec, so it is easily to recognize by name if the pods have the same spec.
+// The additional `postSuffix` can be provided. It can be used to distinguish restarts of POD.
 func CreatePodSuffix(spec api.DeploymentSpec) string {
-	if features.PodNames().Enabled() {
+	if features.ShortPodNames().Enabled() || features.RandomPodNames().Enabled() {
 		return ""
 	}