+++ pkg/deployment/resources/pod_creator_agent_args_test.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2020 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2021 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -84,7 +84,7 @@ func (i inspectorMockStruct) AddService(t *testing.T, svc ...*core.Service) insp
 }
 
 func (i inspectorMockStruct) Get(t *testing.T) inspectorInterface.Inspector {
-	return inspector.NewInspectorFromData(nil, nil, nil, i.services, nil, nil, nil, nil)
+	return inspector.NewInspectorFromData(nil, nil, nil, i.services, nil, nil, nil, nil, nil)
 }
 
 // TestCreateArangodArgsAgent tests createArangodArgs for agent.